requirements.txt
paho-mqtt
pandas
streamlit
sqlite-utils
ortools